@@ -17,10 +17,14 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "axios": "^1.3.4",
-    "eslint-config-next": "13.3.4",
     "framer-motion": "^10.0.1",
-    "next": "13.2.1",
+
+
     "next-seo": "^6.0.0",
+
+    "next": "13.4.6",
+
+
     "nextjs-progressbar": "^0.0.16",
     "plaiceholder": "^2.5.0",
     "react": "18.2.0",
@@ -40,6 +44,7 @@
     "@typescript-eslint/eslint-plugin": "^5.43.0",
     "autoprefixer": "^10.4.13",
     "eslint": "^8.0.1",
+    "eslint-config-next": "13.4.6",
     "eslint-config-prettier": "^8.7.0",
     "eslint-config-standard-with-typescript": "^34.0.1",
     "eslint-plugin-import": "^2.25.2",
