@@ -1,14 +1,15 @@
-import { ISnippet } from '@/types/snippet'
-import { api } from '@/utils/api'
-import { NextPage } from 'next'
+//COMPONENTS
 import Image from 'next/image'
 import Link from 'next/link'
 
-type SnippetsProps = {
-  snippets: ISnippet[]
-}
+//SERVICES
+import { getSnippets } from '@/services/getSnippets'
 
-const Snippets: NextPage<SnippetsProps> = ({ snippets }) => {
+// TYPES
+import { ISnippet } from '@/types/snippet'
+
+export default async function Page() {
+  const snippets: ISnippet[] = await getSnippets('/snippets?populate=*')
   return (
     <div className='grid gap-2 px-6 md:grid-cols-1 md:p-0'>
       <div className='grid gap-5 md:grid-cols-3 md:p-0'>
@@ -48,33 +49,3 @@ const Snippets: NextPage<SnippetsProps> = ({ snippets }) => {
     </div>
   )
 }
-
-export const getStaticProps = async () => {
-  try {
-    const {
-      data: { data: snippets }
-    } = await api.get('/snippets?populate=*')
-
-    const filterSnippets = snippets.map((snippet: ISnippet) => {
-      return {
-        id: snippet.id,
-        attributes: {
-          ...snippet.attributes
-        }
-      }
-    })
-    return {
-      props: {
-        snippets: filterSnippets
-      }
-    }
-  } catch (error) {
-    return {
-      props: {
-        snippets: []
-      }
-    }
-  }
-}
-
-export default Snippets
